last user story complete- refactored playlists into playlists and songs, separated reducers and actions as necessary and updated all react components
