+++ app/javascript/src/containers/SongsIndexContainer.js
@@ -36,7 +36,7 @@ class SongsIndexContainer extends Component {
 
 const mapStateToProps = state => {
   return {
-    artistSongs: state.playlists.artistSongs
+    artistSongs: state.songs.artistSongs
   }
 }
