+++ app/javascript/src/modules/playlists.js
@@ -1,8 +1,6 @@
 import { displayAlertMessage } from './alertMessage.js'
 
 const initialState = {
-  artists: [],
-  artistSongs: [],
   selectedArtistId: null,
   playlistSongs: [],
   isFetching: false
@@ -18,22 +16,6 @@ const playlists = (state = initialState, action) => {
     case REQUEST_FAILED:
       return {
         ...state,
-        isFetching: false
-      }
-    case GET_ARTISTS_REQUEST_SUCCESS:
-      const newArtists = state.artists.concat(action.artists)
-      return {
-        ...state,
-        isFetching: false,
-        artists: newArtists
-      }
-    case GET_ARTIST_SONGS_REQUEST_SUCCESS:
-      const newSongs = action.artistSongs
-      const newArtistId = action.selectedArtistId
-      return {
-        ...state,
-        artistSongs: newSongs,
-        selectedArtistId: newArtistId,
         isFetching: false
       }
     case ADD_SONG_TO_PLAYLIST_SUCCESS:
@@ -80,59 +62,6 @@ const getArtistsRequestSuccess = artists => {
   return {
     type: GET_ARTISTS_REQUEST_SUCCESS,
     artists
-  }
-}
-
-const getArtists = () => {
-  return (dispatch) => {
-    dispatch(startNewRequest())
-    
-    return fetch('/api/v1/artists.json')
-      .then(response => {
-        if (response.ok) {
-          return response.json()
-        } else {
-          dispatch(requestFailed())
-          dispatch(displayAlertMessage("Something went wrong."))
-          return { error: 'Something went wrong' }
-        }
-      })
-      .then(artists => {
-        if (!artists.error) {
-          dispatch(getArtistsRequestSuccess(artists))
-        }
-      })
-  }
-}
-
-const GET_ARTIST_SONGS_REQUEST_SUCCESS = 'GET_ARTIST_SONGS_REQUEST_SUCCESS'
-const getArtistSongsRequestSuccess = (artistId, artistSongs) => {
-  return {
-    type: GET_ARTIST_SONGS_REQUEST_SUCCESS,
-    selectedArtistId: artistId,
-    artistSongs
-  }
-}
-
-const getArtistSongs = (artistId) => {
-  return (dispatch) => {
-    dispatch(startNewRequest())
-    
-    return fetch(`/api/v1/artists/${artistId}/songs`)
-      .then(response => {
-        if (response.ok) {
-          return response.json()
-        } else {
-          dispatch(requestFailed())
-          dispatch(displayAlertMessage("Something went wrong."))
-          return { error: 'Something went wrong' }
-        }
-      })
-      .then(artistSongs => {
-        if (!artistSongs.error) {
-          dispatch(getArtistSongsRequestSuccess(artistId, artistSongs))
-        }
-      })
   }
 }
 
@@ -236,8 +165,6 @@ const removeSongFromPlaylist = (songId) => {
 
 export {
   playlists,
-  getArtists,
-  getArtistSongs,
   addSongToPlaylistPost,
   getExistingPlaylist,
   removeSongFromPlaylist
