+++ app/javascript/src/store/rootReducer.js
@@ -1,10 +1,12 @@
 import { combineReducers }          from 'redux';
 
 import { playlists } from '../modules/playlists'
+import { songs } from '../modules/songs'
 import { alertMessage } from '../modules/alertMessage'
 
 let rootReducer = combineReducers({
   playlists,
+  songs,
   alertMessage
 });
